Moved storage-api to separate branch
Since moving to a storage API might not be as good as just rethinking how 1) we use the allocator system, and 2) how we will handle data layout internally.
It's only being kept on a separate branch to be revisited in the future.
It would also have complicated debugging, as we wouldn't store raw pointers internally anymoer